Fixed: Incorrect fps reported to subtitle renderer. In EVR custom 23.976 FPS was always used in VMR9 25 FPS was always used. Fixed: Handling of different video resolution from video surface resolution. Fixed: Incorrectly reported video resolution and aspect ratio with VMR9 without mixer mode worked around by checking media type. Fixed: Always use 1 pass bicubic resampler to avoid problem of checker pattern caused by different resolution textures used in 2 pass bicubic. Fixed: Incorrect aspect ratio calculation when aspect ratios are very large causing integer overflow. Fixed: Reset device could lock up player in VMR9 renderer mode. (For example when moving between monitors)
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1103 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/subpic/ISubPic.cpp
@@ -1038,7 +1038,7 @@ STDMETHODIMP_(SIZE) ISubPicAllocatorPresenterImpl::GetVideoSize(bool fCorrectAR)
 	CSize VideoSize(m_NativeVideoSize);
 
 	if(fCorrectAR && m_AspectRatio.cx > 0 && m_AspectRatio.cy > 0)
-		VideoSize.cx = VideoSize.cy*m_AspectRatio.cx/m_AspectRatio.cy;
+		VideoSize.cx = (LONGLONG(VideoSize.cy)*LONGLONG(m_AspectRatio.cx))/LONGLONG(m_AspectRatio.cy);
 
 	return(VideoSize);
 }
